+++ matching/src/main/java/com/ssafy/matching/service/GatheringServiceImpl.java
@@ -72,7 +72,7 @@ public class GatheringServiceImpl implements GatheringService {
     }
 
     @Override
-    public void leaveGathering(int gatheringId, int memberId) {
+    public void leaveGathering(int gatheringId, long memberId) {
         memberGatheringRepository.deleteByGatheringIdAndMemberId(gatheringId, memberId);
     }
 }
